close #60

+++ Backend/Api/Controllers/TransactionController.cs
@@ -3,14 +3,14 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Api.Dtos;
 using AutoMapper;
+using Core.Entities.Identity;
 using Core.Interfaces;
 using Core.WebDtos;
-using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Controllers
 {
-    [Authorize]
     public class TransactionController : BaseApiController
     {
         private readonly ITransactionService _transactionService;
@@ -18,7 +18,8 @@ namespace Api.Controllers
 
         public TransactionController(
             ITransactionService transactionService,
-            IMapper mapper)
+            IMapper mapper,
+            UserManager<AppUser> userManager ) : base ( userManager )
         {
             _transactionService = transactionService;
             _mapper = mapper;
@@ -27,6 +28,8 @@ namespace Api.Controllers
         [HttpPost]
         public async Task<ActionResult> CreateTransactionAsync(TransactionWebDto transactionWebDto)
         {
+            if( !IsLogin() ) return Unauthorized();
+            
             var transaction = await _transactionService.CreateTransactionAsync ( transactionWebDto );
 
             if( transaction == null ) return BadRequest();
@@ -35,18 +38,22 @@ namespace Api.Controllers
         }
 
         [HttpGet]
-        public async Task<IReadOnlyList<TransactionDto>> GetTransactionsAsync([FromQuery] DateTime date, string pointName )
+        public async Task<ActionResult<IReadOnlyList<TransactionDto>>> GetTransactionsAsync([FromQuery] DateTime date, string pointName )
         {
+            if( !IsLogin() ) return Unauthorized();
+            
             var transactions = await _transactionService.GetTransactionAsync ( date, pointName );
 
             var transactionsToReturn = _mapper.Map<IReadOnlyList<TransactionDto>> ( transactions );
-            
-            return transactionsToReturn;
+
+            return Ok ( transactionsToReturn );
         }
 
         [HttpPut]
         public async Task<ActionResult<TransactionDto>> UpdateTransactionAsync( TransactionWebDto transactionWebDto )
         {
+            if( !IsLogin() ) return Unauthorized();
+            
             var transaction = await _transactionService.UpdateTransactionAsync ( transactionWebDto );
             
             if (transaction == null) return BadRequest();
@@ -60,6 +67,8 @@ namespace Api.Controllers
         [HttpDelete]
         public async Task<ActionResult> DeleteTransactionAsync( [FromQuery] int id )
         {
+            if( !IsLogin() ) return Unauthorized();
+            
             var result = await _transactionService.DeleteTransactionAsync ( id );
 
             return result > 0 ? Ok() : BadRequest();
@@ -69,6 +78,8 @@ namespace Api.Controllers
         public async Task<ActionResult<IReadOnlyList<TransactionDeletedDto>>> GetDeletedTransactionsAsync 
             ([FromQuery] DateTime? insertedDate, string pointName)
         {
+            if( !await IsAdmin() ) return Unauthorized();
+            
             var deletedTransactions = await _transactionService.GetDeletedTransactionsAsync ( insertedDate, pointName );
             if (deletedTransactions == null) return BadRequest();
 
@@ -81,6 +92,8 @@ namespace Api.Controllers
         [HttpGet ( "product-categories" )]
         public async Task<ActionResult<List<ProductCategoriesToReturnDto>>> GetProductCategoriesAsync()
         {
+            if( !IsLogin() ) return Unauthorized();
+            
             var productCategories = await _transactionService.GetProductCategories();
 
             var productCategoriesToReturn = _mapper.Map<IReadOnlyList<ProductCategoriesToReturnDto>> ( productCategories );
